requirements_mysql.txt
Flask==2.3.3
Flask-Login==0.6.2
Werkzeug==2.3.7
python-dotenv==1.0.0
mysql-connector-python==8.0.33
cryptography==3.4.8